LayoutTests/http/tests/inspector-enabled/contentSecurityPolicy-blocks-setTimeout.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="X-WebKit-CSP" content="script-src 'self' 'unsafe-inline'">
<script src="../inspector/console-test.js"></script>
<script src="../inspector/inspector-test.js"></script>
<script>
function log(text) {
    document.body.appendChild(document.createTextNode(text));
    document.body.appendChild(document.createElement('br'));
}
function onload() {
  if (window.testRunner) {
      testRunner.waitUntilDone();
      testRunner.showWebInspector();
  }

  var id = setTimeout("log('FAIL')", 0);
  if (id != 0)
      log('FAIL: Return value for string (should be 0): ' + id);
  else
      log('PASS: Return value for blocked setTimeout is 0');

  runTest();
}

function test() {
    InspectorTest.dumpConsoleMessages();
    InspectorTest.completeTest();
}
</script>
</head>
<body onload="onload();">
    <p>
        Tests that a CSP-blocked `setTimeout` call generates a console message
        with a stack trace (https://bugs.webkit.org/show_bug.cgi?id=86943). This
        should be followed by a rejection notice tied to line 18.
    </p>
</body>
</html>
